рецепты

+++ src/actions/ingredient.tsx
@@ -54,6 +54,23 @@ export async function getIngredients() {
 
 export async function deleteIngredient(id: string) {
   try {
+    // Проверим, есть ли рецепты, использующие этот ингредиент
+    const recipesUsingIngredient = await prisma.recipe.findMany({
+      where: {
+        ingredients: {
+          some: {
+            ingredientId: id
+          }
+        }
+      }
+    });
+
+    if (recipesUsingIngredient.length > 0) {
+      throw new Error(
+        `Ингредиент используется в ${recipesUsingIngredient.length} рецептах`
+      );
+    }
+
     const ingredient = await prisma.ingredient.delete({
       where: { id }
     });
